requirements.txt
Authlib==1.2.1
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pypdf2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
stripe==7.0.0
typing-extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.3.0
wtforms==3.1.0
zipp==3.17.0
